Add files via upload
Income is often a useful indicator of a society's quality of life and wealth. 
Household income is a commonly cited measure of individual wealth. For e.g. Economists use household income to draw a host of conclusions about the economic health of a given area or population. Per capita income measures the average income earned by each person in a given area.
Monitoring HH income is important for government, donors, researchers, and others because increasing rural household income is a primary objective for achieving many development goals, including reducing poverty, hunger, and food and nutrition insecurity. So, we here as a team tried to develop a machine learning model to predict the income.